:recycle: Move delete_all_element_version_data_files to versioning module
