+++ containers/daap-athena-load/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [1.2.4] 2023-11.6
+## [1.2.4] 2023-11-6
 
 ### Changed
